+++ src/Libraries/SmartStore.Data/Migrations/201506261756463_PrimaryStoreCurrencyMultiStore.cs
@@ -131,6 +131,12 @@ namespace SmartStore.Data.Migrations
 				"Specifies whether the store should be SSL secured.",
 				"Legt fest, ob der Shop SSL gesichert werden soll.");
 
+			builder.AddOrUpdate("Admin.Configuration.Settings.News.MaxAgeInDays",
+				"Maximum age (in days)",
+				"Maximales Alter (in Tagen)",
+				"Specifies the maximum news age in days. Older news are not exported in the RSS feed.",
+				"Legt das maximale News-Alter in Tagen fest. Ältere News werden im RSS-Feed nicht exportiert.");
+
 			builder.Delete("Admin.Configuration.Currencies.CantDeletePrimary");
 			builder.Delete("Admin.Configuration.Currencies.CantDeleteExchange");
 			builder.Delete("Admin.Configuration.Currencies.Fields.MarkAsPrimaryStoreCurrency");